# Hoist all functions to the top level