--- a/src/execution.ts
+++ b/src/execution.ts
@@ -178,6 +178,10 @@ export interface CompiledQuery {
   stringify: (v: any) => string;
 }
 
+interface InternalCompiledQuery extends CompiledQuery {
+  __DO_NOT_USE_THIS_OR_YOU_WILL_BE_FIRED_compilation: string;
+}
+
 /**
  * It compiles a GraphQL query to an executable function
  * @param {GraphQLSchema} schema GraphQL schema
@@ -237,7 +241,7 @@ export function compileQuery(
     );
 
     const functionBody = compileOperation(context);
-    return {
+    const compiledQuery: InternalCompiledQuery = {
       query: createBoundQuery(
         context,
         document,
@@ -247,8 +251,12 @@ export function compileQuery(
           ? context.operation.name.value
           : undefined
       ),
-      stringify
+      stringify,
+      // result of the compilation useful for debugging issues
+      // and visualization tools like try-jit.
+      __DO_NOT_USE_THIS_OR_YOU_WILL_BE_FIRED_compilation: functionBody
     };
+    return compiledQuery;
   } catch (err) {
     return {
       errors: normalizeErrors(err)