--- a/src/variables.ts
+++ b/src/variables.ts
@@ -1,6 +1,7 @@
 import genFn from "generate-function";
 import {
   GraphQLBoolean,
+  GraphQLError,
   GraphQLFloat,
   GraphQLID,
   GraphQLInputType,
@@ -18,12 +19,25 @@ import {
   valueFromAST,
   VariableDefinitionNode
 } from "graphql";
-import { CoercedVariableValues } from "graphql/execution/values";
 import { addPath, computeLocations, ObjectPath } from "./ast";
-import { GraphQLError } from "./error";
+import { GraphQLError as GraphJITQLError } from "./error";
 import createInspect from "./inspect";
 
 const inspect = createInspect();
+
+export type CoercedVariableValues = FailedVariableCoertion | VariableValues;
+
+interface FailedVariableCoertion {
+  errors: ReadonlyArray<GraphQLError>;
+}
+
+interface VariableValues {
+  coerced: { [key: string]: any };
+}
+
+export function failToParseVariables(x: any): x is FailedVariableCoertion {
+  return x.errors;
+}
 
 interface CompilationContext {
   inputPath: ObjectPath;
@@ -62,7 +76,7 @@ export function compileVariableParsing(
       // Must use input types for variables. This should be caught during
       // validation, however is checked again here for safety.
       errors.push(
-        new (GraphQLError as any)(
+        new (GraphJITQLError as any)(
           `Variable "$${varName}" expected value of type ` +
             `"${
               varType ? varType : print(varDefNode.type)
@@ -106,12 +120,12 @@ export function compileVariableParsing(
 
   return Function.apply(
     null,
-    ["GraphQLError", "inspect"]
+    ["GraphJITQLError", "inspect"]
       .concat(Array.from(dependencies.keys()))
       .concat(gen.toString())
   ).apply(
     null,
-    [GraphQLError, inspect].concat(Array.from(dependencies.values()))
+    [GraphJITQLError, inspect].concat(Array.from(dependencies.values()))
   );
 }
 
@@ -154,7 +168,7 @@ function generateInput(
     varType = varType.ofType;
     gen(`
       if (${currentOutput} == null) {
-        errors.push(new GraphQLError(${hasValueName} ? ${nonNullMessage} : ${omittedMessage}, ${errorLocation}));
+        errors.push(new GraphJITQLError(${hasValueName} ? ${nonNullMessage} : ${omittedMessage}, ${errorLocation}));
       }
     `);
   } else {
@@ -172,7 +186,7 @@ function generateInput(
           } else if (Number.isInteger(${currentInput})) {
             ${currentOutput} = ${currentInput}.toString();
           } else {
-            errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+            errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
               inspect(${currentInput}) + "; " +
               'Expected type ${varType.name}; ' +
               '${varType.name} cannot represent value: ' +
@@ -186,7 +200,7 @@ function generateInput(
           if (typeof ${currentInput} === "string") {
               ${currentOutput} = ${currentInput};
           } else {
-            errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+            errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
               inspect(${currentInput}) + "; " +
               'Expected type ${varType.name}; ' +
               '${varType.name} cannot represent a non string value: ' +
@@ -200,7 +214,7 @@ function generateInput(
         if (typeof ${currentInput} === "boolean") {
             ${currentOutput} = ${currentInput};
         } else {
-          errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+          errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
           inspect(${currentInput}) + "; " +
           'Expected type ${varType.name}; ' +
           '${varType.name} cannot represent a non boolean value: ' +
@@ -212,7 +226,7 @@ function generateInput(
         gen(`
         if (Number.isInteger(${currentInput})) {
           if (${currentInput} > ${MAX_32BIT_INT} || ${currentInput} < ${MIN_32BIT_INT}) {
-            errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+            errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
             inspect(${currentInput}) + "; " +
             'Expected type ${varType.name}; ' +
             '${
@@ -223,7 +237,7 @@ function generateInput(
             ${currentOutput} = ${currentInput};
           }
         } else {
-          errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+          errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
             inspect(${currentInput}) + "; " +
             'Expected type ${varType.name}; ' +
             '${varType.name} cannot represent non-integer value: ' +
@@ -237,7 +251,7 @@ function generateInput(
         if (Number.isFinite(${currentInput})) {
             ${currentOutput} = ${currentInput};
         } else {
-          errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+          errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
             inspect(${currentInput}) + "; " +
             'Expected type ${varType.name}; ' +
             '${varType.name} cannot represent non numeric value: ' +
@@ -255,13 +269,13 @@ function generateInput(
           try {
             const parseResult = ${varType.name}parseValue(${currentInput});
             if (parseResult === undefined || parseResult !== parseResult) {
-              errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+              errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
               inspect(${currentInput}) + "; " +
               'Expected type ${varType.name}.', ${errorLocation}));
             }
             ${currentOutput} = parseResult;
           } catch (error) {
-            errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+            errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
               inspect(${currentInput}) + "; " +
               'Expected type ${varType.name}.', ${errorLocation})
             );
@@ -280,14 +294,14 @@ function generateInput(
           ${currentOutput} = enumValue.value;
         } else {
           errors.push(
-            new GraphQLError('Variable "$${varName}" got invalid value ' +
+            new GraphJITQLError('Variable "$${varName}" got invalid value ' +
             inspect(${currentInput}) + "; " +
             'Expected type ${varType.name}.', ${errorLocation})
           );
         }
       } else {
         errors.push(
-          new GraphQLError('Variable "$${varName}" got invalid value ' +
+          new GraphJITQLError('Variable "$${varName}" got invalid value ' +
           inspect(${currentInput}) + "; " +
           'Expected type ${varType.name}.', ${errorLocation})
         );
@@ -327,7 +341,7 @@ function generateInput(
   } else if (isInputType(varType)) {
     gen(`
       if (typeof ${currentInput} !== 'object') {
-        errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+        errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
         inspect(${currentInput}) + "; " +
         'Expected type ${varType.name} to be an object.', ${errorLocation}));
       } else {
@@ -362,7 +376,7 @@ function generateInput(
       const allowedFields = ${JSON.stringify(allowedFields)};
       for (const fieldName of Object.keys(${currentInput})) {
         if (!allowedFields.includes(fieldName)) {
-          errors.push(new GraphQLError('Variable "$${varName}" got invalid value ' +
+          errors.push(new GraphJITQLError('Variable "$${varName}" got invalid value ' +
             inspect(${currentInput}) + "; " +
             'Field "' + fieldName + '" is not defined by type ${
               varType.name