Currently, graphql-jit creates a lot of dynamic functions and takes advantages of closures to propagate context related with errors.

This creates additional memory pressure that could be avoided by hoisting the functions.
The PR introduces an execution context that is passed around to avoid the multiple closures well as
hoisting the grand majority of function created. The only ones left are the ones related with the abstract types which will be improved in a later PR.

Master
```
$ ts-node -T ./src/__benchmarks__/benchmarks.ts
Starting introspection
graphql-jit x 2,280 ops/sec ±5.12% (208 runs sampled)
Starting fewResolvers
graphql-jit x 97,191 ops/sec ±1.96% (222 runs sampled)
Starting manyResolvers
graphql-jit x 56,322 ops/sec ±4.60% (218 runs sampled)
Starting nestedArrays
graphql-jit x 485 ops/sec ±2.13% (219 runs sampled)
```

This PR
```
$ ts-node -T ./src/__benchmarks__/benchmarks.ts
Starting introspection
graphql-jit x 4,525 ops/sec ±1.04% (226 runs sampled)
Starting fewResolvers
graphql-jit x 121,675 ops/sec ±1.06% (221 runs sampled)
Starting manyResolvers
graphql-jit x 100,029 ops/sec ±1.38% (221 runs sampled)
Starting nestedArrays
graphql-jit x 915 ops/sec ±2.49% (222 runs sampled)
```